# swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Notes management service",
        "title": "Notes Service.",
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/info": {
            "get": {
                "description": "Get service information",
                "tags": [
                    "info"
                ],
                "summary": "Get service information",
                "operationId": "info-status",
                "security": [
                    {
                        "my_auth": [
                            "email"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/InfoResponse"
                    }
                }
            }
        },
        "/notes": {
            "get": {
                "description": "Returns the entire list of notes",
                "tags": [
                    "notes"
                ],
                "summary": "Fetches all the notes",
                "operationId": "notes-fetchAll",
                "security": [
                    {
                        "my_auth": [
                            "email"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/FetchAllNotesResponse"
                    }
                }
            },
            "post": {
                "description": "Adds a new note",
                "tags": [
                    "notes"
                ],
                "summary": "Add a new note",
                "operationId": "notes-add",
                "security": [
                    {
                        "my_auth": [
                            "email"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AddNoteRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/Success"
                    }
                }
            }
        },
        "/notes/{id}": {
            "get": {
                "description": "Fetch a single note by index",
                "tags": [
                    "notes"
                ],
                "summary": "Fetch a single note",
                "operationId": "notes-fetchOne",
                "security": [
                    {
                        "my_auth": [
                            "email"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "x-go-name": "NoteId",
                        "description": "Index of Note",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/FetchNoteResponse"
                    }
                }
            },
            "put": {
                "description": "Updates the note at the specified index",
                "tags": [
                    "notes"
                ],
                "summary": "Updates a single note",
                "operationId": "notes-update",
                "security": [
                    {
                        "my_auth": [
                            "email"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "x-go-name": "NoteId",
                        "description": "Index of Note",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AddNoteRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/Success"
                    }
                }
            },
            "delete": {
                "description": "Removes a single note by index",
                "tags": [
                    "notes"
                ],
                "summary": "Removes a single note",
                "operationId": "notes-remove",
                "security": [
                    {
                        "my_auth": [
                            "email"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int64",
                        "x-go-name": "NoteId",
                        "description": "Index of Note",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/Success"
                    }
                }
            }
        }
    },
    "definitions": {
        "AddNoteRequestBody": {
            "description": "AddNoteRequestBody",
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "x-go-name": "Content"
                },
                "title": {
                    "type": "string",
                    "x-go-name": "Title"
                }
            },
            "x-go-package": "notes/app"
        },
        "FetchAllNotesResponseBody": {
            "type": "object",
            "properties": {
                "notes": {
                    "description": "The Note List",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Note"
                    },
                    "x-go-name": "Notes"
                }
            },
            "x-go-package": "notes/app"
        },
        "Info": {
            "type": "object",
            "properties": {
                "service": {
                    "type": "string",
                    "x-go-name": "Service"
                },
                "status": {
                    "type": "string",
                    "x-go-name": "Status"
                }
            },
            "x-go-package": "notes/app"
        },
        "Note": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "x-go-name": "Content"
                },
                "title": {
                    "type": "string",
                    "x-go-name": "Title"
                }
            },
            "x-go-package": "notes/app"
        }
    },
    "responses": {
        "FetchAllNotesResponse": {
            "description": "The response containing the list of Notes",
            "schema": {
                "$ref": "#/definitions/FetchAllNotesResponseBody"
            }
        },
        "FetchNoteResponse": {
            "description": "The response for a single note fetch",
            "schema": {
                "$ref": "#/definitions/Note"
            }
        },
        "InfoResponse": {
            "description": "InfoResponse",
            "schema": {
                "$ref": "#/definitions/Info"
            }
        },
        "Success": {
            "description": "Success"
        }
    },
    "securityDefinitions": {
        "my_auth": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://accounts.google.com/o/oauth2/v2/auth",
            "scopes": {
                "email": "default scope"
            }
        }
    }
}